Risolto bug scroll dei tab. Adesso se si preme back sulla main activity, previo pop up, viene richiesto se effettuare il log out. Spostato il codice per il log out dai vari fragment alla shop activity.

+++ Client/app/src/main/java/com/example/cocktailapp/Carrello.java
@@ -59,8 +59,8 @@ public class Carrello {
 
 
     // Calcola il totale del carrello
-    public float calculateTotal() {
-        float total = 0;
+    public double calculateTotal() {
+        double total = 0;
 
         // Calcola il totale dei cocktail
         for (Cocktail cocktail : cocktails) {
